+++ src/SDL/Input/Joystick.hs
@@ -15,6 +15,7 @@ module SDL.Input.Joystick
 
   , getJoystickID
   , Joystick
+  , JoyButtonState(..)
   , buttonPressed
   , ballDelta
   , axisPosition
@@ -57,6 +58,16 @@ data JoystickDevice = JoystickDevice
   { joystickDeviceName :: Text
   , joystickDeviceId :: CInt
   } deriving (Eq, Generic, Read, Ord, Show, Typeable)
+
+-- | Identifies the state of a joystick button.
+data JoyButtonState = JoyButtonPressed | JoyButtonReleased
+  deriving (Data, Eq, Generic, Ord, Read, Show, Typeable)
+
+instance FromNumber JoyButtonState Word8 where
+  fromNumber n = case n of
+    Raw.SDL_PRESSED -> JoyButtonPressed
+    Raw.SDL_RELEASED -> JoyButtonReleased
+    _ -> JoyButtonReleased
 
 -- | Count the number of joysticks attached to the system.
 --